requirements.txt
agency-swarm
mem0ai
